First version of ASP.NET code (#10)
* Initial code - RESTful api

* [backend] Adds Event model, basic CRUD and Swagger

Adds the Event model and controller, with basic CRUD operations. Also,
configures Swagger to the project, in order to provide
self-documentation features to our project.

* [backend] Adding user model and event foreign key

Adds the user model and controller for basic CRUD operations. Events now
have a foreign key, connecting them to a user (this is for testing
only, will be changed later). Also adds necessary packages for EF with PostgreSQL.

* [backend] Now fully using ORM for data persistency

* README for code and study-material

Adds the initial READMEs for code, which breafly explains what is inside
the directory. Also, updating links used for study in
study-material/README.md.

* Some naming changes to the project

* Adding status meeting

Moves the sprint1-planning to the correct folder, and also adds the
status meeting for today in the project.

* Changing launchUrl - now opens Swagger